requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0
huggingface-hub>=0.20.2
python-multipart>=0.0.6
aiohttp>=3.9.1
numpy>=1.26.3

